Make file names and exports more consistent
Use `createX` pattern for all modules that export factory
functions.

Export factory function as default for all modules that
export factory functions.

Move query interface to interfaces file, to separate it
from the createCompoundQuery factory.
