+++ src/query/createStoreRange.ts
@@ -1,12 +1,14 @@
-import { Query, QueryType } from './createQuery';
-interface StoreRange<T> extends Query<T, T> {
+import { Query, QueryType } from './interfaces';
+export interface StoreRange<T> extends Query<T, T> {
 	readonly start: number;
 	readonly count: number;
 }
 
-export default StoreRange;
+function serializeRange(range: StoreRange<any>): string {
+	return `range(${range.start}, ${range.count})`;
+}
 
-export function createRange<T>(start: number, count: number, serializer?: (range: StoreRange<T>) => string): StoreRange<T> {
+function createRange<T>(start: number, count: number, serializer?: (range: StoreRange<T>) => string): StoreRange<T> {
 	return {
 		apply(data: T[]) {
 			return data.slice(start, start + count);
@@ -21,6 +23,4 @@ export function createRange<T>(start: number, count: number, serializer?: (range
 	};
 }
 
-function serializeRange(range: StoreRange<any>): string {
-	return `range(${range.start}, ${range.count})`;
-}
+export default createRange;
