+++ src/store/createSubcollectionStore.ts
@@ -1,10 +1,9 @@
-import createStore, { StoreOptions, CrudOptions, Store, PatchArgument } from './createStore';
-import { UpdateResults } from '../storage/createInMemoryStorage';
+import createStore, { StoreOptions, CrudOptions, Store, PatchArgument, UpdateResults } from './createStore';
 import { ComposeFactory } from 'dojo-compose/compose';
 import Promise from 'dojo-shim/Promise';
 import WeakMap from 'dojo-shim/WeakMap';
 import { StoreObservable } from './createStoreObservable';
-import { Query } from '../query/createQuery';
+import { Query } from '../query/interfaces';
 import { mixin } from 'dojo-core/lang';
 
 export interface SubcollectionOptions<T, O extends CrudOptions, U extends UpdateResults<T>> extends StoreOptions<T, O> {
