+++ tests/unit/store/createStore.ts
@@ -2,17 +2,16 @@ import * as registerSuite from 'intern!object';
 import * as assert from 'intern/chai!assert';
 import * as sinon from 'sinon';
 import createStore, { StoreOperation } from '../../../src/store/createStore';
-import Patch from '../../../src/patch/Patch';
 import Map from 'dojo-shim/Map';
 import Set from 'dojo-shim/Set';
 import Promise from 'dojo-shim/Promise';
-import { createRange } from '../../../src/query/StoreRange';
-import { createFilter } from '../../../src/query/Filter';
-import { createPointer } from '../../../src/patch/JsonPointer';
-import { createPatch } from '../../../src/patch/Patch';
-import { createSort } from '../../../src/query/Sort';
-import { createOperation, OperationType } from '../../../src/patch/Operation';
-import createCompoundQuery from '../../../src/query/createQuery';
+import createRange from '../../../src/query/createStoreRange';
+import createFilter from '../../../src/query/createFilter';
+import createPointer from '../../../src/patch/createJsonPointer';
+import createPatch, { Patch } from '../../../src/patch/createPatch';
+import createSort from '../../../src/query/createSort';
+import createOperation, { OperationType } from '../../../src/patch/createOperation';
+import createCompoundQuery from '../../../src/query/createCompoundQuery';
 import createInMemoryStorage from '../../../src/storage/createInMemoryStorage';
 import { createData, ItemType, createUpdates, patches, patchedItems } from '../support/createData';
 import createAsyncStorage from '../support/AsyncStorage';
